Refactor format-toolbar code to use split web/native files
