+++ packages/block-editor/src/components/rich-text/format-toolbar-container.native.js
@@ -1,0 +1,16 @@
+/**
+ * Internal dependencies
+ */
+import BlockFormatControls from '../block-format-controls';
+import FormatToolbar from './format-toolbar';
+
+const FormatToolbarContainer = () => {
+	// Render regular toolbar
+	return (
+		<BlockFormatControls>
+			<FormatToolbar />
+		</BlockFormatControls>
+	);
+};
+
+export default FormatToolbarContainer;
